+++ src/main/java/net/openhft/chronicle/bytes/algo/OptimisedBytesStoreHash.java
@@ -114,7 +114,7 @@ public enum OptimisedBytesStoreHash implements BytesStoreHash<BytesStore> {
     public static long applyAsLong32bytesMultiple(@NotNull BytesStore store, int remaining) throws BufferUnderflowException {
         @NotNull final BytesStore bytesStore = store.bytesStore();
         final long address = bytesStore.addressForRead(store.readPosition());
-        long h0 = remaining * K0, h1 = 0, h2 = 0, h3 = 0;
+        long h0 = (long) remaining * K0, h1 = 0, h2 = 0, h3 = 0;
 
         int i;
         for (i = 0; i < remaining - 31; i += 32) {
